Fixed early memory free; Added free_... functions to make fortify checking easier; Added fortify scopes; Delete now moves to trash can on systems with the xworkplace trash can installed.

+++ dll/eas.c
@@ -1144,7 +1144,7 @@ PVOID SaveEA(CHAR * filename, HOLDFEA * current, CHAR * newdata,
       }
       break;
     }
-    pfealist->list[0].cbValue = (ULONG) (eaval -
+    pfealist->list[0].cbValue = /*(ULONG)*/ (eaval -
 					 (pfealist->list[0].szName +
 					  pfealist->list[0].cbName + 1));
     eaop.fpGEA2List = (PGEA2LIST) 0;
